* h2 support: make PR compile
