+++ quill-jdbc/src/test/resources/application.conf
@@ -13,5 +13,5 @@ testPostgresDB.dataSource.portNumber=${?POSTGRES_PORT_5432_TCP_PORT}
 testPostgresDB.dataSource.serverName=${?POSTGRES_PORT_5432_TCP_ADDR}
 
 testH2DB.dataSourceClassName=org.h2.jdbcx.JdbcDataSource
-testH2DB.dataSource.url=jdbc:h2:mem:test;DB_CLOSE_DELAY=-1;INIT=RUNSCRIPT FROM 'classpath:sql/h2-schema.sql'
+testH2DB.dataSource.url="jdbc:h2:mem:test;DB_CLOSE_DELAY=-1;INIT=RUNSCRIPT FROM 'classpath:sql/h2-schema.sql'"
 testH2DB.dataSource.user=sa
