+++ quill-jdbc/src/test/scala/io/getquill/sources/jdbc/h2/JdbcEncodingSpec.scala
@@ -1,13 +1,13 @@
-package io.getquill.source.jdbc.h2
+package io.getquill.sources.jdbc.h2
 
 import io.getquill._
-import io.getquill.source.sql.EncodingSpec
+import io.getquill.sources.sql.EncodingSpec
 
 class JdbcEncodingSpec extends EncodingSpec {
 
   "encodes and decodes types" in {
     testH2DB.run(delete)
-    testH2DB.run(insert).using(insertValues)
+    testH2DB.run(insert)(insertValues)
     verify(testH2DB.run(query[EncodingTestEntity]))
   }
 }
