+++ quill-jdbc/src/test/scala/io/getquill/sources/jdbc/h2/PeopleJdbcSpec.scala
@@ -1,10 +1,7 @@
-package io.getquill.source.jdbc.h2
+package io.getquill.sources.jdbc.h2
 
 import io.getquill._
-import io.getquill.source.sql.PeopleSpec
-import io.getquill.quotation.Quoted
-import io.getquill.quotation.IsDynamic
-import io.getquill.source.sql.mirror.mirrorSource
+import io.getquill.sources.sql.PeopleSpec
 
 class PeopleJdbcSpec extends PeopleSpec {
 
@@ -12,8 +9,8 @@ class PeopleJdbcSpec extends PeopleSpec {
     val t = testH2DB.transaction {
       testH2DB.run(query[Couple].delete)
       testH2DB.run(query[Person].filter(_.age > 0).delete)
-      testH2DB.run(peopleInsert).using(peopleEntries)
-      testH2DB.run(couplesInsert).using(couplesEntries)
+      testH2DB.run(peopleInsert)(peopleEntries)
+      testH2DB.run(couplesInsert)(couplesEntries)
     }
   }
 
@@ -22,7 +19,7 @@ class PeopleJdbcSpec extends PeopleSpec {
   }
 
   "Example 2 - range simple" in {
-    testH2DB.run(`Ex 2 rangeSimple`).using(`Ex 2 param 1`, `Ex 2 param 2`) mustEqual `Ex 2 expected result`
+    testH2DB.run(`Ex 2 rangeSimple`)(`Ex 2 param 1`, `Ex 2 param 2`) mustEqual `Ex 2 expected result`
   }
 
   "Example 3 - satisfies" in {
@@ -34,7 +31,7 @@ class PeopleJdbcSpec extends PeopleSpec {
   }
 
   "Example 5 - compose" in {
-    testH2DB.run(`Ex 5 compose`).using(`Ex 5 param 1`, `Ex 5 param 2`) mustEqual `Ex 5 expected result`
+    testH2DB.run(`Ex 5 compose`)(`Ex 5 param 1`, `Ex 5 param 2`) mustEqual `Ex 5 expected result`
   }
 
   "Example 6 - predicate 0" in {
