+++ quill-sql/src/test/sql/mysql-schema.sql
@@ -1,30 +1,30 @@
 CREATE TABLE Person(
-	name VARCHAR(255),
-	age int(3)
+    name VARCHAR(255),
+    age int(3)
 );
 
 CREATE TABLE Couple(
-	her VARCHAR(255),
-	him VARCHAR(255)
+    her VARCHAR(255),
+    him VARCHAR(255)
 );
 
 CREATE TABLE Department(
-	dpt VARCHAR(255)
+    dpt VARCHAR(255)
 );
 
 CREATE TABLE Employee(
-	emp VARCHAR(255),
-	dpt VARCHAR(255),
-	salary int(20)
+    emp VARCHAR(255),
+    dpt VARCHAR(255),
+    salary int(20)
 );
 
 CREATE TABLE Task(
-	emp VARCHAR(255),
-	tsk VARCHAR(255)
+    emp VARCHAR(255),
+    tsk VARCHAR(255)
 );
 
 CREATE TABLE EncodingTestEntity(
-	v1 VARCHAR(255),
+    v1 VARCHAR(255),
     v2 DECIMAL(5,2),
     v3 BOOLEAN,
     v4 SMALLINT,
@@ -49,26 +49,26 @@ CREATE TABLE EncodingTestEntity(
 );
 
 Create TABLE DateEncodingTestEntity(
-  v1 date,
-  v2 datetime,
-  v3 timestamp
+    v1 date,
+    v2 datetime,
+    v3 timestamp
 );
 
 CREATE TABLE TestEntity(
-	s VARCHAR(255),
+    s VARCHAR(255),
     i INTEGER,
     l BIGINT,
     o INTEGER
 );
 
 CREATE TABLE TestEntity2(
-	s VARCHAR(255),
+    s VARCHAR(255),
     i INTEGER,
     l BIGINT
 );
 
 CREATE TABLE TestEntity3(
-	s VARCHAR(255),
+    s VARCHAR(255),
     i INTEGER,
     l BIGINT
 );
